apps/api/requirements.txt
# ResumeAI API — Python 3.12

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
httpx==0.27.2

# AI
anthropic==0.34.2

# Cache
redis[asyncio]==5.1.0

# File parsing
pdfplumber==0.11.4
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Export
weasyprint==62.3
jinja2==3.1.4

# Billing
stripe==10.12.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0
